refactor: restructure chat scraping logic into modular components
Split the monolithic scraping.ts into separate utility modules for browser initialization, message processing, and chat scraping. This improves maintainability and separation of concerns.

- Created browser.ts for Puppeteer setup
- Added messageProcessor.ts for message filtering and saving
- Introduced chatScraper.ts for chat message extraction
- Updated index.ts to use new modules
- Cleaned up tsconfig.json formatting
